+++ Game/src/SpacialPartitioning/Entity.cs
@@ -5,11 +5,11 @@ namespace SpacialPartitioning;
 public class Entity
 {
     public Vector2 Position;
-    public ulong Id;
+    public uint Id;
     public float Radius { get; private set; }
     public float RadiusSqr { get; private set; }
 
-    public Entity(Vector2 pos, ulong id, float radius)
+    public Entity(Vector2 pos, uint id, float radius)
     {
         Position = pos;
         Id = id;
